handle_satori/requirements.txt
functions-framework==3.5.0
flask==2.3.3
requests==2.31.0
google-cloud-tasks==2.19.2
google-cloud-logging==3.9.0
google-auth==2.29.0
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.15.0
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pydantic==2.6.1
python-dotenv==1.0.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pandas==2.2.0
openpyxl==3.1.2